Changed R wrapper to reduce peak memory usage.

+++ trunk/src/AT_SPC.c
@@ -21,7 +21,7 @@
  *    GNU General Public License for more details.
  *
  *    You should have received a copy of the GNU General Public License
- *    along with AmTrack (file: copying.txt).
+ *    aint with AmTrack (file: copying.txt).
  *    If not, see <http://www.gnu.org/licenses/>
  */
 
@@ -44,7 +44,7 @@ void endian_swap4(unsigned int* x)
 }
 
 
-// __int64 for MSVC, "long long" for gcc
+// __int64 for MSVC, "int int" for gcc
 void endian_swap8(uint64_t * x)
 {
   *x = (*x>>56) |
@@ -72,20 +72,20 @@ void readStruct(FILE *fp,
 }
 
 
-long AT_SPC_get_size_from_filename(const char filename[FILE_NAME_NCHAR])
+int AT_SPC_get_size_from_filename(const char filename[FILE_NAME_NCHAR])
 {
 	FILE* fp;
 	fp = fopen(filename, "rb");
-    long size = AT_SPC_get_size(fp);
+    int size = AT_SPC_get_size(fp);
     fclose(fp);
 	return(size);
 }
 
 
-long AT_SPC_get_size(FILE *fp){
+int AT_SPC_get_size(FILE *fp){
 	char string[80];
 	bool switchEndian=false;
-	long size=0;
+	int size=0;
 	struct STRPSPCBTAG  filetype;
 	readStruct(fp,&filetype,false);
 	//Skip structure. Just going to assume the tags are right here.
@@ -200,27 +200,27 @@ long AT_SPC_get_size(FILE *fp){
 }
 
 
-long AT_SPC_read(const char filename[FILE_NAME_NCHAR],
-		long* depth_step[],
+int AT_SPC_read(const char filename[FILE_NAME_NCHAR],
+		int* depth_step[],
 		double* depth_g_cm2[],
 		double* E_MeV_u[],
 		double* DE_MeV_u[],
-		long* particle_no[],
+		int* particle_no[],
 		double* fluence_cm2[]){
 
 	FILE* fp;
 	fp = fopen(filename, "rb");
 
-	long total_n_bins = AT_SPC_get_size(fp);
+	int total_n_bins = AT_SPC_get_size(fp);
 
-	*depth_step    = (long*)realloc(*depth_step, total_n_bins * sizeof(long));
+	*depth_step    = (int*)realloc(*depth_step, total_n_bins * sizeof(int));
 	*depth_g_cm2   = (double*)realloc(*depth_g_cm2, total_n_bins * sizeof(double));
 	*E_MeV_u       = (double*)realloc(*E_MeV_u, total_n_bins * sizeof(double));
 	*DE_MeV_u      = (double*)realloc(*DE_MeV_u, total_n_bins * sizeof(double));
-	*particle_no   = (long*)realloc(*particle_no, total_n_bins * sizeof(long));
+	*particle_no   = (int*)realloc(*particle_no, total_n_bins * sizeof(int));
 	*fluence_cm2   = (double*)realloc(*fluence_cm2, total_n_bins * sizeof(double));
 
-	long n_bins_read = AT_SPC_read_data(fp,
+	int n_bins_read = AT_SPC_read_data(fp,
 			total_n_bins,
 			*depth_step,
 			*depth_g_cm2,
@@ -235,19 +235,19 @@ long AT_SPC_read(const char filename[FILE_NAME_NCHAR],
 }
 
 
-long AT_SPC_read_data_from_filename( const char filename[FILE_NAME_NCHAR],
-		const long n,
-		long depth_step[],
+int AT_SPC_read_data_from_filename( const char filename[FILE_NAME_NCHAR],
+		const int n,
+		int depth_step[],
 		double depth_g_cm2[],
 		double E_MeV_u[],
 		double DE_MeV_u[],
-		long particle_no[],
+		int particle_no[],
 		double fluence_cm2[])
 {
 	FILE* fp;
 	fp = fopen(filename, "rb");
 
-	long n_bins_read = AT_SPC_read_data(fp,
+	int n_bins_read = AT_SPC_read_data(fp,
 			n,
 			depth_step,
 			depth_g_cm2,
@@ -262,18 +262,18 @@ long AT_SPC_read_data_from_filename( const char filename[FILE_NAME_NCHAR],
 }
 
 
-long AT_SPC_read_data(FILE* fp,
-		const long n,
-		long depth_step[],
+int AT_SPC_read_data(FILE* fp,
+		const int n,
+		int depth_step[],
 		double depth_g_cm2[],
 		double E_MeV_u[],
 		double DE_MeV_u[],
-		long particle_no[],
+		int particle_no[],
 		double fluence_cm2[]){
 
 	char string[80];
 	bool switchEndian=false;
-	long size=0;
+	int size=0;
 	unsigned int i=0;
 	unsigned int k =0;
 	int particles[1000];
